+++ crates/deadpool/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add unique `id` to objects which can be read via the `Object::id` method
+- Add `WeakPool<T>` and `Pool::weak()` for non-owning, upgradeable pool references.
 
 ## [0.12.2] - 2025-02-02